% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateAbbrToName.R
\name{stateAbbrToName}
\alias{stateAbbrToName}
\title{Convert state abbreviations to state names}
\usage{
stateAbbrToName(df)
}
\arguments{
\item{df}{A \code{data.frame}, expected to contain at least two variables:
\code{StateName}, and \code{FIPS}.}
}
\value{
The same \code{data.frame}, but with all abbreviations found in
\code{StateName} or \code{FIPS} replaced with full names.
}
\description{
This makes use of Arul John's \code{us-states} repository to convert state
abbrevations/territory abbreviations to state/territory names
}
