% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccineAdjust-package.R
\docType{package}
\name{vaccineAdjust-package}
\alias{vaccineAdjust}
\alias{vaccineAdjust-package}
\title{vaccineAdjust: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fayette Klaassen \email{klaassen.fayette@gmail.com} (\href{https://orcid.org/0000-0002-9488-2571}{ORCID})

}
\keyword{internal}
